requirements.txt
wheel
python-chess
pytest
pytest-cov
twine
pypandoc
